pyproject.toml
[project]
name = "image-to-tex"
version = "0.1.0"
description = "Convert images of scientific work (equations, tables, diagrams) to LaTeX code using vision AI models"
readme = "README.md"
authors = [
    { name = "anjor", email = "anjor@umd.edu" }
]
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "click>=8.1.7",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
image-to-tex = "image_to_tex.cli.commands:cli"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
